+++ instrumentation/github.com/labstack/echo/otelecho/echo.go
@@ -28,9 +28,7 @@ const (
 
 // Middleware returns echo middleware which will trace incoming requests.
 func Middleware(serverName string, opts ...Option) echo.MiddlewareFunc {
-	cfg := config{
-		HandleError: true,
-	}
+	cfg := config{}
 	for _, opt := range opts {
 		opt.apply(&cfg)
 	}
@@ -95,9 +93,6 @@ func Middleware(serverName string, opts ...Option) echo.MiddlewareFunc {
 			err := next(c)
 			if err != nil {
 				span.SetAttributes(attribute.String("echo.error", err.Error()))
-				if cfg.HandleError {
-					c.Error(err)
-				}
 			}
 
 			status := c.Response().Status